move fci_mo to sci
