+++ src/forte_old_methods.cc
@@ -53,7 +53,7 @@
 #include "fci/fci.h"
 #include "fci/fci_solver.h"
 #include "fci/fci_integrals.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "finite_temperature.h"
 #include "helpers.h"
 #include "sci/mrci.h"
