+++ src/mrdsrg-helper/run_dsrg.h
@@ -2,7 +2,7 @@
 #define _run_dsrg_h_
 
 #include "forte_options.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "fci/fci.h"
 #include "mrdsrg-spin-integrated/dsrg_mrpt2.h"
 #include "mrdsrg-spin-integrated/dsrg_mrpt3.h"
