+++ src/mrdsrg-spin-integrated/dsrg_mrpt2.cc
@@ -41,7 +41,7 @@
 #include "ci_rdm/ci_rdms.h"
 #include "fci/fci.h"
 #include "boost/format.hpp"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "fci/fci_solver.h"
 #include "helpers/printing.h"
 #include "dsrg_mrpt2.h"
