+++ src/mrdsrg-spin-integrated/dsrg_mrpt2_ms.cc
@@ -31,7 +31,7 @@
 #include "helpers/timer.h"
 #include "ci_rdm/ci_rdms.h"
 #include "fci/fci_solver.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "dsrg_mrpt2.h"
 
 namespace psi {
