+++ src/mrdsrg-spin-integrated/dsrg_mrpt3.cc
@@ -42,7 +42,7 @@
 #include "helpers/timer.h"
 #include "helpers/blockedtensorfactory.h"
 #include "fci/fci.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "boost/format.hpp"
 #include "helpers/printing.h"
 #include "dsrg_mrpt3.h"
