+++ src/mrdsrg-spin-integrated/dwms_mrpt2.cc
@@ -9,7 +9,7 @@
 #include "dsrg_mrpt2.h"
 #include "dsrg_mrpt3.h"
 #include "three_dsrg_mrpt2.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "orbital-helpers/semi_canonicalize.h"
 #include "helpers/printing.h"
