+++ src/mrdsrg-spin-integrated/mrdsrg.cc
@@ -34,7 +34,7 @@
 #include "psi4/libmints/molecule.h"
 
 #include "fci/fci.h"
-#include "fci_mo.h"
+#include "sci/fci_mo.h"
 #include "helpers/printing.h"
 #include "orbital-helpers/semi_canonicalize.h"
 #include "boost/format.hpp"
