+++ src/sci/fci_mo.cc
